requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
yfinance>=0.2.40
matplotlib>=3.9.0
pandas>=2.2.0
python-dotenv==1.0.0
